backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
faiss-cpu
numpy
google-genai
